requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
requests
beautifulsoup4
playwright
spacy
nltk
sentence-transformers
faiss-cpu
streamlit
